--- reactbits/dock.jsx
@@ -1,210 +1,0 @@
-'use client';
-
-import {
-  Children,
-  cloneElement,
-  useEffect,
-  useMemo,
-  useRef,
-  useState,
-} from 'react';
-
-import {
-  AnimatePresence,
-  motion,
-  useMotionValue,
-  useSpring,
-  useTransform,
-} from 'framer-motion';
-
-import '../src/styles/Editor/dock.css';
-import ZenMode from './ZenMode';
-
-function DockItem({
-  children,
-  className = '',
-  onClick,
-  mouseX,
-  spring,
-  distance,
-  magnification,
-  baseItemSize,
-}) {
-  const ref = useRef(null);
-  const isHovered = useMotionValue(0);
-
-  const mouseDistance = useTransform(mouseX, (val) => {
-    const rect = ref.current?.getBoundingClientRect() ?? {
-      x: 0,
-      width: baseItemSize,
-    };
-    return val - rect.x - baseItemSize / 2;
-  });
-
-  const targetSize = useTransform(
-    mouseDistance,
-    [-distance, 0, distance],
-    [baseItemSize, magnification, baseItemSize],
-  );
-  const size = useSpring(targetSize, spring);
-
-  // Enhanced click handler with debugging
-  const handleClick = (e) => {
-    e.preventDefault();
-    e.stopPropagation();
-    console.log('Dock item clicked!');
-    if (typeof onClick === 'function') {
-      onClick();
-    }
-  };
-
-  return (
-    <motion.div
-      ref={ref}
-      style={{
-        width: size,
-        height: size,
-      }}
-      onHoverStart={() => isHovered.set(1)}
-      onHoverEnd={() => isHovered.set(0)}
-      onFocus={() => isHovered.set(1)}
-      onBlur={() => isHovered.set(0)}
-      onClick={handleClick}
-      className={`dock-item ${className}`}
-      tabIndex={0}
-      role="button"
-      aria-haspopup="true"
-    >
-      {Children.map(children, (child) => cloneElement(child, { isHovered }))}
-    </motion.div>
-  );
-}
-
-function DockLabel({ children, className = '', ...rest }) {
-  const { isHovered } = rest;
-  const [isVisible, setIsVisible] = useState(false);
-
-  useEffect(() => {
-    const unsubscribe = isHovered.on('change', (latest) => {
-      setIsVisible(latest === 1);
-    });
-    return () => unsubscribe();
-  }, [isHovered]);
-
-  return (
-    <AnimatePresence>
-      {isVisible && (
-        <motion.div
-          initial={{ opacity: 0, y: 0 }}
-          animate={{ opacity: 1, y: -10 }}
-          exit={{ opacity: 0, y: 0 }}
-          transition={{ duration: 0.2 }}
-          className={`dock-label ${className}`}
-          role="tooltip"
-          style={{ x: '-50%' }}
-        >
-          {children}
-        </motion.div>
-      )}
-    </AnimatePresence>
-  );
-}
-
-function DockIcon({ children, className = '' }) {
-  return <div className={`dock-icon ${className}`}>{children}</div>;
-}
-
-export default function Dock({
-  items,
-  className = '',
-  spring = { mass: 0.1, stiffness: 150, damping: 12 },
-  magnification = 70,
-  distance = 200,
-  panelHeight = 68,
-  dockHeight = 256,
-  baseItemSize = 50,
-}) {
-  const mouseX = useMotionValue(Infinity);
-  const isHovered = useMotionValue(0);
-  const [zenModeActive, setZenModeActive] = useState(false);
-
-  const maxHeight = useMemo(
-    () => Math.max(dockHeight, magnification + magnification / 2 + 4),
-    [magnification, dockHeight],
-  );
-  const heightRow = useTransform(isHovered, [0, 1], [panelHeight, maxHeight]);
-  const height = useSpring(heightRow, spring);
-
-  // Toggle Zen Mode
-  const toggleZenMode = () => {
-    console.log('Toggling Zen Mode, current state:', !zenModeActive);
-    setZenModeActive(!zenModeActive);
-  };
-
-  // Add a Zen Mode item if not already in the items array
-  const dockItems = useMemo(() => {
-    // Check if a Zen Mode item already exists in the items array
-    const zenModeExists = items.some(item => 
-      item.label && item.label.toLowerCase().includes('zen')
-    );
-
-    if (!zenModeExists) {
-      // Create a new array with the zen mode item added
-      return [
-        ...items,
-        {
-          icon: '🌧️',
-          label: 'Zen Mode',
-          onClick: toggleZenMode,
-          className: 'zen-mode-item',
-        },
-      ];
-    }
-    
-    return items;
-  }, [items]);
-
-  return (
-    <>
-      <motion.div
-        style={{ height, scrollbarWidth: 'none' }}
-        className="dock-outer"
-      >
-        <motion.div
-          onMouseMove={({ pageX }) => {
-            isHovered.set(1);
-            mouseX.set(pageX);
-          }}
-          onMouseLeave={() => {
-            isHovered.set(0);
-            mouseX.set(Infinity);
-          }}
-          className={`dock-panel ${className}`}
-          role="toolbar"
-          aria-label="Application dock"
-        >
-          {dockItems.map((item, index) => (
-            <DockItem
-              key={index}
-              onClick={item.onClick}
-              className={item.className}
-              mouseX={mouseX}
-              spring={spring}
-              distance={distance}
-              magnification={magnification}
-              baseItemSize={baseItemSize}
-            >
-              <DockIcon>{item.icon}</DockIcon>
-              <DockLabel>{item.label}</DockLabel>
-            </DockItem>
-          ))}
-        </motion.div>
-      </motion.div>
-      
-      <ZenMode 
-        isActive={zenModeActive} 
-        onClose={() => setZenModeActive(false)} 
-      />
-    </>
-  );
-}
